Integration tests breakdown: Extract integration_core_hosts_test
Relates to #51839

Move 21 host lifecycle/listing test methods (TestListHosts, TestHostsCount,
TestBulkDeleteHosts*, TestHostsAddToTeam, TestHostDeviceMapping*, etc.) out of
integration_core_test.go into their own file, with a header comment describing
what belongs there. Pure move, no test logic changed.
